services/jwt_issuer/requirements.txt
fastapi==0.115.4
uvicorn==0.32.0
python-jose==3.3.0
